Create full_financials_shares_prices
I extended the example given to incorporate all 3 financial statements (income, balance, cashflow) as well as sharecount and share price data as a single Excel sheet download. I fooled around a bit to make the code determine how much data was available and give you up to 8 quarters of data and all the annual data available, as well as sharecount and price data for those quarters where it's available. 
I'm not a practiced coder, so any input or comments would be very welcome